make util file for page.tsx export function
